requirements.txt
boto3==1.26.137
botocore==1.29.137
Flask==2.3.2
python-dotenv==1.0.0
pytest==7.3.2
pytest-cov==4.1.0
moto==4.1.10
Pillow==10.0.0
